Pulled metatag generation into a new method
